src/server/pyproject.toml
# server/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "graphcap-server"
version = "0.0.2"
description = "Server component for graphcap "
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.9.0",
    "asyncpg>=0.30.0",
    "fastapi>=0.115.8",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.7.1",
    "sqlalchemy[asyncio]>=2.0.37",
    "uvicorn>=0.34.0",
    "dagster-postgres>=0.25.11",
    "dagster-graphql>=1.9.11",
    "loguru>=0.7.3",
    "tenacity>=9.0.0",
    "python-multipart>=0.0.20",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["server"]

[tool.hatch.build]
packages = ["server"]

